+++ domain/proc/BlameableBehavior.php
@@ -1,0 +1,64 @@
+<?php
+
+namespace domain\proc;
+
+use Yii;
+use yii\behaviors\AttributeBehavior;
+use yii\db\BaseActiveRecord;
+
+class BlameableBehavior extends AttributeBehavior
+{
+    /**
+     * @var string the attribute that will receive current user ID value
+     * Set this property to false if you do not want to record the creator ID.
+     */
+    public $createdByAttribute = 'created_by';
+    /**
+     * @var string the attribute that will receive current user ID value
+     * Set this property to false if you do not want to record the updater ID.
+     */
+    public $updatedByAttribute = 'updated_by';
+    /**
+     * @inheritdoc
+     *
+     * In case, when the property is `null`, the value of `Yii::$app->user->id` will be used as the value.
+     */
+    public $value;
+
+
+    /**
+     * @inheritdoc
+     */
+    public function init()
+    {
+        parent::init();
+
+        if (empty($this->attributes)) {
+            $this->attributes = [
+                BaseActiveRecord::EVENT_BEFORE_INSERT => [$this->createdByAttribute, $this->updatedByAttribute],
+                BaseActiveRecord::EVENT_BEFORE_UPDATE => $this->updatedByAttribute,
+            ];
+        }
+    }
+
+    /**
+     * @inheritdoc
+     *
+     * In case, when the [[value]] property is `null`, the value of `Yii::$app->user->id` will be used as the value.
+     */
+    protected function getValue($event)
+    {
+        if ($this->value === null) {
+            $user = Yii::$app->get('user', false);
+            $session = Yii::$app->get('session', false);
+
+            if ($user && $session) {
+                return $user->isGuest ? $session->id : $user->id;
+            }
+
+            return null;
+        }
+
+        return parent::getValue($event);
+    }
+}