+++ domain/proc/models/ReportLoader.php
@@ -2,6 +2,7 @@
 
 namespace domain\proc\models;
 
+use domain\proc\BlameableBehavior;
 use Yii;
 use yii\behaviors\TimestampBehavior;
 use yii\db\Expression;
@@ -14,6 +15,7 @@ use yii\db\Expression;
  * @property string $rl_report_id
  * @property string $rl_report_filename
  * @property string $rl_report_displayname
+ * @property string $rl_report_type
  * @property integer $rl_status
  * @property integer $rl_percent
  * @property integer $rl_start
@@ -34,10 +36,12 @@ class ReportLoader extends \yii\db\ActiveRecord
     public function rules()
     {
         return [
-            [['rl_process_id', 'rl_report_id', 'rl_report_filename', 'rl_report_displayname'], 'required'],
+            [['rl_process_id', 'rl_report_id', 'rl_report_filename', 'rl_report_displayname', 'rl_report_type'], 'required'],
             [['rl_status', 'rl_percent', 'rl_start'], 'integer'],
             [['rl_process_id', 'rl_report_id'], 'string', 'max' => 64],
             [['rl_report_filename', 'rl_report_displayname'], 'string', 'max' => 255],
+            [['rl_report_type'], 'string', 'max' => 10],
+            [['rl_report_filename', 'default', 'value' => Yii::getAlias('@common') . '/tmpfiles/report' . time() . '.xlsx']],
         ];
     }
 
@@ -47,18 +51,24 @@ class ReportLoader extends \yii\db\ActiveRecord
             [
                 'class' => TimestampBehavior::className(),
                 'createdAtAttribute' => 'rl_start',
+                'updatedAtAttribute' => false,
                 'value' => new Expression('NOW()'),
+            ],
+            [
+                'class' => BlameableBehavior::className(),
+                'createdAtAttribute' => 'rl_process_id',
+                'updatedAtAttribute' => false,
             ],
         ];
     }
 
-    public function beforeSave($insert)
-    {
-        if ($insert) {
-            $this->rl_process_id = Yii::$app->user->isGuest ? Yii::$app->session->getId() : Yii::$app->user->getId();
-            $this->rl_report_filename = Yii::getAlias('@common') . '/tmpfiles/report' . time();
-        }
+    /*  public function beforeSave($insert)
+   {
+    if ($insert) {
+           $this->rl_process_id = Yii::$app->user->isGuest ? Yii::$app->session->getId() : Yii::$app->user->getId();
+           $this->rl_report_filename = Yii::getAlias('@common') . '/tmpfiles/report' . time() . '.xlsx';
+       }
 
-        return parent::beforeSave($insert);
-    }
+       return parent::beforeSave($insert);
+   }*/
 }